changes to DBLoader to add units and changes to test
